+++ src/game/Mechanic.type.ts
